+++ src/components/generate/GenerateView/FlashcardReviewSection/FlashcardReviewSection.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 import { Button } from "@/components/ui/button";
 import type { FlashcardProposalViewModel } from "../../../../types";
 import { FlashcardList } from "../../../FlashcardList";
-import { useFlashcardReview } from "./useFlashcardReview";
+import { useFlashcardReview } from "../hooks";
 
 interface FlashcardReviewSectionProps {
   flashcards: FlashcardProposalViewModel[];
