+++ src/components/generate/GenerateView/hooks/useFlashcardStats/index.ts
@@ -1,0 +1,1 @@
+export { useFlashcardStats, type FlashcardStats } from "./useFlashcardStats";
